+++ src/test/java/org/janelia/saalfeldlab/n5/metadata/NgffTests.java
@@ -5,7 +5,8 @@ import static org.junit.Assert.fail;
 import java.io.IOException;
 
 import org.janelia.saalfeldlab.n5.N5FSReader;
-import org.janelia.saalfeldlab.n5.metadata.NgffMultiScaleGroupAttributes.MultiscaleDataset;
+import org.janelia.saalfeldlab.n5.universe.metadata.NgffMultiScaleGroupAttributes;
+import org.janelia.saalfeldlab.n5.universe.metadata.NgffMultiScaleGroupAttributes.MultiscaleDataset;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
